pyproject.toml
[project]
name = "weather-app"
version = "0.1.0"
description = ""
authors = [
    {name = "iShmuel",email = "123bberger@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.47.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai (==0.28)",
    "folium (>=0.20.0,<0.21.0)",
    "streamlit-folium (>=0.25.0,<0.26.0)"
]

[tool.poetry]
packages = [{include = "weather_app", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

